requirements.txt
streamlit
pandas
python-bcb
yfinance
plotly